Move `CoreDataFeedStore` implementation of the `FeedImageDataStore` protocol to production
